+++ src/utils/settings.py
@@ -111,7 +111,7 @@ APPLICATION = "commix"
 DESCRIPTION_FULL = "Automated All-in-One OS Command Injection and Exploitation Tool"
 DESCRIPTION = "The command injection exploiter"
 AUTHOR  = "Anastasios Stasinopoulos"
-VERSION_NUM = "2.5.11"
+VERSION_NUM = "2.5.12"
 STABLE_VERSION = False
 if STABLE_VERSION:
   VERSION = "v" + VERSION_NUM[:3] + "-stable"
@@ -680,7 +680,8 @@ TAMPER_SCRIPTS = {
                   "space2vtab": False,
                   "singlequotes": False,
                   "caret": False,
-                  "multiplespaces":False
+                  "multiplespaces":False,
+                  "backslashes":False
                  }
 
 # HTTP Errors