+++ readme/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Version 2.5 (upcoming)
+* Added: New tamper script "backslashes.py" that adds back slashes (\) between the characters of the generated payloads (for *nix targets).
 * Fixed: Minor bug-fix regarding unicode decode exception error due to invalid codec, during connection on target host.
 * Revised: Improvement regarding combining tamper script "multiplespaces.py" with other space-related tamper script(s).
 * Added: New tamper script "multiplespaces.py" that adds multiple spaces around OS commands.